+++ spec/requests/carto/api/legends_controller_spec.rb
@@ -138,9 +138,9 @@ module Carto
       end
 
       describe '#create' do
-        before(:each) { @layer.reload.legends.map(&:destroy) }
+        after(:each) { @layer.reload.legends.map(&:destroy) }
 
-        def create_lengend_url(user: @user, visualization: @visualization, layer: @layer)
+        def create_legend_url(user: @user, visualization: @visualization, layer: @layer)
           legends_url(user_domain: user.subdomain,
                       visualization_id: visualization.id,
                       layer_id: layer.id,
@@ -148,7 +148,7 @@ module Carto
         end
 
         it 'should reject non visualization owners' do
-          post_json create_lengend_url(user: @intruder), {} do |response|
+          post_json create_legend_url(user: @intruder), {} do |response|
             response.status.should eq 403
           end
         end
@@ -156,31 +156,15 @@ module Carto
         it 'should reject non data layers' do
           base_layer = @visualization.layers.first
 
-          post_json create_lengend_url(layer: base_layer), custom_legend_payload do |response|
+          post_json create_legend_url(layer: base_layer), custom_legend_payload do |response|
             response.status.should eq 422
 
             response.body[:errors].should include("'#{base_layer.kind}' layers can't have legends")
           end
         end
 
-        it 'should reject more that Legend:MAX_LEGENDS_PER_LAYER legends per layer' do
-          Legend::MAX_LEGENDS_PER_LAYER.times do
-            post_json create_lengend_url, custom_legend_payload do |response|
-              response.status.should eq 201
-
-              legend_is_correct(response.body)
-            end
-          end
-
-          post_json create_lengend_url, custom_legend_payload do |response|
-            response.status.should eq 422
-
-            response.body[:errors].should include('Maximum number of legends per layer reached')
-          end
-        end
-
         it 'should return a validation errors' do
-          post_json create_lengend_url, {} do |response|
+          post_json create_legend_url, {} do |response|
             response.status.should eq 422
 
             response.body[:errors].should_not be_nil
@@ -189,7 +173,7 @@ module Carto
 
         describe 'with valid definitions' do
           after(:each) do
-            post_json create_lengend_url, @payload do |response|
+            post_json create_legend_url, @payload do |response|
               response.status.should eq 201
 
               legend_is_correct(response.body)
@@ -226,7 +210,7 @@ module Carto
             spammy_definition = @payload[:definition].merge(spam: 'hell')
             @payload[:definition] = spammy_definition
 
-            post_json create_lengend_url, @payload do |response|
+            post_json create_legend_url, @payload do |response|
               response.status.should eq 422
             end
           end
@@ -261,7 +245,7 @@ module Carto
             spammy_definition = @payload[:conf].merge(spam: 'hell')
             @payload[:conf] = spammy_definition
 
-            post_json create_lengend_url, @payload do |response|
+            post_json create_legend_url, @payload do |response|
               response.status.should eq 422
             end
           end
@@ -295,7 +279,7 @@ module Carto
           after(:each) do
             @payload[:conf] = { columns: { not: 'an_array' } }
 
-            post_json create_lengend_url, @payload do |response|
+            post_json create_legend_url, @payload do |response|
               response.status.should eq 422
             end
           end
@@ -329,7 +313,7 @@ module Carto
           after(:each) do
             @payload[:conf] = ['manolo', 'escobar', 2]
 
-            post_json create_lengend_url, @payload do |response|
+            post_json create_legend_url, @payload do |response|
               response.status.should eq 422
             end
           end
@@ -381,13 +365,49 @@ module Carto
             response.body[:errors].should include('Visualization not found')
           end
         end
+
+        describe 'multiple legends per layer' do
+          it 'can create a color layer and a size layer' do
+            post_json create_legend_url, category_legend_payload do |response|
+              response.status.should eq 201
+            end
+
+            post_json create_legend_url, bubble_legend_payload do |response|
+              response.status.should eq 201
+            end
+          end
+
+          it 'cannot create two color layers' do
+            post_json create_legend_url, category_legend_payload do |response|
+              response.status.should eq 201
+            end
+
+            post_json create_legend_url, custom_legend_payload do |response|
+              response.status.should eq 422
+
+              response.body[:errors].should include('Only one color legend per layer allowed')
+            end
+          end
+
+          it 'cannot create two size layers' do
+            post_json create_legend_url, bubble_legend_payload do |response|
+              response.status.should eq 201
+            end
+
+            post_json create_legend_url, bubble_legend_payload do |response|
+              response.status.should eq 422
+
+              response.body[:errors].should include('Only one size legend per layer allowed')
+            end
+          end
+        end
       end
 
       describe '#show' do
         before (:all) { @legend = Legend.create!(custom_legend_payload.merge(layer_id: @layer.id)) }
         after  (:all) { @legend.destroy }
 
-        def show_lengend_url(user: @user, visualization: @visualization, layer: @layer, legend: @legend)
+        def show_legend_url(user: @user, visualization: @visualization, layer: @layer, legend: @legend)
           legend_url(user_domain: user.subdomain,
                      visualization_id: visualization.id,
                      layer_id: layer.id,
@@ -396,7 +416,7 @@ module Carto
         end
 
         it 'should show a legend' do
-          get_json show_lengend_url, {} do |response|
+          get_json show_legend_url, {} do |response|
             response.status.should eq 200
 
             legend_is_correct(response.body)
@@ -404,7 +424,7 @@ module Carto
         end
 
         it 'should not show a legend to others' do
-          get_json show_lengend_url(user: @intruder), {} do |response|
+          get_json show_legend_url(user: @intruder), {} do |response|
             response.status.should eq 403
           end
         end
@@ -427,7 +447,7 @@ module Carto
         before (:all) { @legend = Legend.create!(custom_legend_payload.merge(layer_id: @layer.id)) }
         after  (:all) { @legend.destroy }
 
-        def update_lengend_url(user: @user, visualization: @visualization, layer: @layer, legend: @legend)
+        def update_legend_url(user: @user, visualization: @visualization, layer: @layer, legend: @legend)
           legend_url(user_domain: user.subdomain,
                      visualization_id: visualization.id,
                      layer_id: layer.id,
@@ -437,7 +457,7 @@ module Carto
 
         it 'updates a legend' do
           custom_legend_payload[:definition][:html] = '<p>modified</p>'
-          put_json update_lengend_url, custom_legend_payload do |response|
+          put_json update_legend_url, custom_legend_payload do |response|
             response.status.should eq 200
 
             legend_is_correct(response.body)
@@ -445,12 +465,10 @@ module Carto
         end
 
         it 'updates a legend when max legends reached' do
-          (Legend::MAX_LEGENDS_PER_LAYER - 1).times do
-            Legend.create!(custom_legend_payload.merge(layer_id: @layer.id))
-          end
+          Legend.create!(bubble_legend_payload.merge(layer_id: @layer.id))
 
           custom_legend_payload[:definition][:html] = '<p>modified</p>'
-          put_json update_lengend_url, custom_legend_payload do |response|
+          put_json update_legend_url, custom_legend_payload do |response|
             response.status.should eq 200
 
             legend_is_correct(response.body)
@@ -462,7 +480,7 @@ module Carto
         end
 
         it 'should let others update a legend' do
-          put_json update_lengend_url(user: @intruder), {} do |response|
+          put_json update_legend_url(user: @intruder), {} do |response|
             response.status.should eq 403
           end
         end
@@ -485,14 +503,14 @@ module Carto
         before(:all) do
           @layer.reload.legends.map(&:destroy)
           Legend.create!(custom_legend_payload.merge(layer_id: @layer.id))
-          Legend.create!(custom_legend_payload.merge(layer_id: @layer.id))
+          Legend.create!(bubble_legend_payload.merge(layer_id: @layer.id))
         end
 
         after(:all) do
           @layer.reload.legends.map(&:destroy)
         end
 
-        def index_lengend_url(user: @user, visualization: @visualization, layer: @layer)
+        def index_legend_url(user: @user, visualization: @visualization, layer: @layer)
           legends_url(user_domain: user.subdomain,
                       visualization_id: visualization.id,
                       layer_id: layer.id,
@@ -500,7 +518,7 @@ module Carto
         end
 
         it 'indexes legends' do
-          get_json index_lengend_url, {} do |response|
+          get_json index_legend_url, {} do |response|
             response.status.should eq 200
 
             response.body.each do |legend|
@@ -510,7 +528,7 @@ module Carto
         end
 
         it 'should not index legends to others' do
-          get_json index_lengend_url(user: @intruder), {} do |response|
+          get_json index_legend_url(user: @intruder), {} do |response|
             response.status.should eq 403
           end
         end
@@ -520,7 +538,7 @@ module Carto
         before (:each) { @legend = Legend.create!(custom_legend_payload.merge(layer_id: @layer.id)) }
         after  (:each) { @legend.destroy }
 
-        def delete_lengend_url(user: @user, visualization: @visualization, layer: @layer, legend: @legend)
+        def delete_legend_url(user: @user, visualization: @visualization, layer: @layer, legend: @legend)
           legend_url(user_domain: user.subdomain,
                      visualization_id: visualization.id,
                      layer_id: layer.id,
@@ -529,7 +547,7 @@ module Carto
         end
 
         it 'should delete a legend' do
-          delete_json delete_lengend_url, {} do |response|
+          delete_json delete_legend_url, {} do |response|
             response.status.should eq 204
 
             response.body.should be_empty
@@ -539,7 +557,7 @@ module Carto
         end
 
         it 'should not delete another user\'s legend' do
-          delete_json delete_lengend_url(user: @intruder), {} do |response|
+          delete_json delete_legend_url(user: @intruder), {} do |response|
             response.status.should eq 403
           end